fix clang compilation

+++ ain_imager_src/imagerwindow.cpp
@@ -599,7 +599,7 @@ void ImagerWindow::on_create_preview(indigo_property *property, indigo_item *ite
 			m_indigo_item = nullptr;
 		}
 		m_indigo_item = item;
-		const stretch_config_t sconfig = {conf.preview_stretch_level, conf.preview_color_balance};
+		const stretch_config_t sconfig = {(uint8_t)conf.preview_stretch_level, (uint8_t)conf.preview_color_balance};
 		preview_cache.create(property, m_indigo_item, sconfig);
 		QString key = preview_cache.create_key(property, m_indigo_item);
 		//preview_image *image = preview_cache.get(m_image_key);
@@ -612,7 +612,7 @@ void ImagerWindow::on_create_preview(indigo_property *property, indigo_item *ite
 	} else if (get_selected_guider_agent(selected_agent)) {
 		if ((client_match_device_property(property, selected_agent, CCD_IMAGE_PROPERTY_NAME) && conf.guider_save_bandwidth == 0) ||
 			(client_match_device_property(property, selected_agent, CCD_PREVIEW_IMAGE_PROPERTY_NAME) && conf.guider_save_bandwidth > 0)) {
-			const stretch_config_t sconfig = {conf.guider_stretch_level, conf.guider_color_balance};
+			const stretch_config_t sconfig = {(uint8_t)conf.guider_stretch_level, (uint8_t)conf.guider_color_balance};
 			preview_cache.create(property, item, sconfig);
 			QString key = preview_cache.create_key(property, item);
 			//preview_image *image = preview_cache.get(key);
@@ -850,7 +850,7 @@ void ImagerWindow::on_use_system_locale_changed(bool status) {
 
 void ImagerWindow::on_imager_stretch_changed(int level) {
 	conf.preview_stretch_level = (preview_stretch)level;
-	const stretch_config_t sc = {conf.preview_stretch_level, conf.preview_color_balance};
+	const stretch_config_t sc = {(uint8_t)conf.preview_stretch_level, (uint8_t)conf.preview_color_balance};
 	preview_cache.recreate(m_image_key, sc);
 	show_preview_in_imager_viewer(m_image_key);
 	write_conf();
@@ -859,7 +859,7 @@ void ImagerWindow::on_imager_stretch_changed(int level) {
 
 void ImagerWindow::on_imager_cb_changed(int balance) {
 	conf.preview_color_balance = (color_balance)balance;
-	const stretch_config_t sc = {conf.preview_stretch_level, conf.preview_color_balance};
+	const stretch_config_t sc = {(uint8_t)conf.preview_stretch_level, (uint8_t)conf.preview_color_balance};
 	preview_cache.recreate(m_image_key, sc);
 	show_preview_in_imager_viewer(m_image_key);
 	write_conf();
@@ -873,7 +873,7 @@ void ImagerWindow::on_guider_stretch_changed(int level) {
 		get_selected_guider_agent(selected_agent);
 		setup_preview(selected_agent);
 	});
-	const stretch_config_t sc = {conf.guider_stretch_level, conf.guider_color_balance};
+	const stretch_config_t sc = {(uint8_t)conf.guider_stretch_level, (uint8_t)conf.guider_color_balance};
 	preview_cache.recreate(m_guider_key, sc);
 	show_preview_in_guider_viewer(m_guider_key);
 	write_conf();
@@ -882,7 +882,7 @@ void ImagerWindow::on_guider_stretch_changed(int level) {
 
 void ImagerWindow::on_guider_cb_changed(int balance) {
 	conf.guider_color_balance = (color_balance)balance;
-	const stretch_config_t sc = {conf.guider_stretch_level, conf.guider_color_balance};
+	const stretch_config_t sc = {(uint8_t)conf.guider_stretch_level, (uint8_t)conf.guider_color_balance};
 	preview_cache.recreate(m_guider_key, sc);
 	show_preview_in_guider_viewer(m_guider_key);
 	write_conf();
